Move InventoryTabs into property folder
