+++ source/frontend/src/features/mapSideBar/property/InventoryTabs.tsx
