+++ source/frontend/src/features/properties/map/acquisition/AcquisitionContainer.tsx
@@ -1,8 +1,8 @@
 import LoadingBackdrop from 'components/common/LoadingBackdrop';
 import { useMapSearch } from 'components/maps/hooks/useMapSearch';
 import { FileTypes } from 'constants/index';
+import { InventoryTabNames } from 'features/mapSideBar/property/InventoryTabs';
 import { FileTabType } from 'features/mapSideBar/tabs/FileTabs';
-import { InventoryTabNames } from 'features/mapSideBar/tabs/InventoryTabs';
 import { FormikProps } from 'formik';
 import { useAcquisitionProvider } from 'hooks/repositories/useAcquisitionProvider';
 import useApiUserOverride from 'hooks/useApiUserOverride';
