+++ source/frontend/src/features/properties/map/acquisition/ViewSelector.tsx
@@ -1,9 +1,9 @@
 import { FileTypes } from 'constants/fileTypes';
+import { InventoryTabNames, InventoryTabs } from 'features/mapSideBar/property/InventoryTabs';
 import { UpdatePropertyDetailsContainer } from 'features/mapSideBar/property/tabs/propertyDetails/update/UpdatePropertyDetailsContainer';
 import TakesUpdateContainer from 'features/mapSideBar/property/tabs/takes/update/TakesUpdateContainer';
 import TakesUpdateForm from 'features/mapSideBar/property/tabs/takes/update/TakesUpdateForm';
 import { FileTabType } from 'features/mapSideBar/tabs/FileTabs';
-import { InventoryTabNames, InventoryTabs } from 'features/mapSideBar/tabs/InventoryTabs';
 import { FormikProps } from 'formik';
 import { Api_AcquisitionFile } from 'models/api/AcquisitionFile';
 import React from 'react';
