+++ source/frontend/src/features/properties/map/shared/detail/PropertyFileContainer.test.tsx
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
 import LoadingBackdrop from 'components/common/LoadingBackdrop';
-import { IInventoryTabsProps, InventoryTabNames } from 'features/mapSideBar/tabs/InventoryTabs';
+import { IInventoryTabsProps, InventoryTabNames } from 'features/mapSideBar/property/InventoryTabs';
 import { noop } from 'lodash';
 import { mockLtsaResponse, mockWfsGetPropertyById } from 'mocks/index.mock';
 import { mockLookups } from 'mocks/lookups.mock';
